Moved files to their respective folders
