+++ C/Remove_whiteSpace_from_string.c
