+++ C/decimaltobinary.c
